Add fetures and corrct bugs

+++ app/src/main/java/com/example/inventariosapp/domain/model/product/ProductIdResponseModel.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class ProductIdResponseModel(
     @SerializedName("productoId") val productoId: Int? = null,
-    @SerializedName("inventario") val inventario: Int? = null,
+    @SerializedName("inventario") val inventario: Double? = null,
 )
